Scripts/emissary_panda_registry/requirements.txt
winregistry
python-registry